tcc-extragalactic-open-set/pyproject.toml
[tool.poetry]
name = "tcc-extragalactic-open-set"
version = "0.1.0"
description = "Pipeline reprodutível para classificação morfológica extragaláctica com open-set."
authors = ["Codex <codex@example.com>"]
readme = "README.md"
packages = [{ include = "tccastro", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
astropy = "^6.1.0"
astroquery = "^0.4.7"
photutils = "^1.13.0"
numpy = "^1.26.4"
pandas = "^2.2.2"
requests = "^2.32.3"
tqdm = "^4.66.4"
scikit-learn = "^1.5.1"
torch = "^2.3.1"
torchvision = "^0.18.1"
matplotlib = "^3.9.0"
typer = "^0.12.3"
pyyaml = "^6.0.1"
joblib = "^1.4.2"
pillow = "^10.4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
ruff = "^0.5.5"
black = "^24.4.2"
mypy = "^1.10.0"
types-requests = "^2.32.0.20240602"

[tool.poetry.scripts]
tccastro = "tccastro.cli:app"

[tool.black]
line-length = 88

target-version = ["py311"]

[tool.ruff]
line-length = 88
select = ["E", "F", "W", "I"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
strict = false

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"
